+++ ANI-KUTA/APP/ani-kuta/core/content/src/main/java/com/confused/anikuta/core/content/ContentRepository.kt
@@ -362,11 +362,11 @@ class ContentRepository(
             return
         }
         val now = System.currentTimeMillis()
-        val order = libraryQueries.countDefaultCategoryItems().executeAsOne().toInt()
+        val order = libraryQueries.countDefaultCategoryItems().executeAsOne()
         libraryQueries.addToDefaultCategory(
             mainId = mainId,
             categoryId = defaultCat.id,
-            displayOrder = order.toLong(),
+            displayOrder = order,
             addedAt = now,
         )
         Logger.i(TAG) { "Added to library (Default): mainId=$mainId" }
